Implemented hard-coded times/stops and myTaskParams
myTaskParams contains the vectors needed for storing stops and stop time
info, and current location.  These can be passed in as one object to my
AsyncTask.

The schedule is hard coded but using algorithms following schedule
patterns to help automate the process.

If there is time in the future, this should be replaced with an xml
parser.  After spending 10+ hours trying to correctly parse the xml,
hard-coding became the most realistic option.