Running with Notifications
Implemented a test location and test time (current time + 2 minutes).
This allows me to test the app logic and the notifications.

Notification implemented, needs refining (message).

App builds, runs, notifies and vibrates.  However, updates don't run
when the app isn't in the foreground (because onPause() is called).

Need to find way to run it in the background yet control when it starts
and stops.