Got rid of myTaskParams copy constructor
It was useless.  All it was used for was making a new copy of the
location in the params[0] for findLocation so that I could manipulate it
without affecting the actual data.

Instead I used Vector<Location> temp = new Vector<>(params[0].location).
This achieves the same purpose with far less code.